Move Acknowledgeable to more appropriate package
