+++ graylog2-server/src/main/java/org/graylog2/plugin/Message.java
@@ -36,11 +36,11 @@ import org.apache.commons.lang3.StringUtils;
 import org.graylog.failure.FailureCause;
 import org.graylog.failure.ProcessingFailureCause;
 import org.graylog2.indexer.IndexSet;
-import org.graylog2.indexer.messages.Acknowledgeable;
 import org.graylog2.indexer.messages.Indexable;
 import org.graylog2.plugin.streams.Stream;
 import org.graylog2.plugin.utilities.date.DateTimeConverter;
 import org.graylog2.plugin.utilities.ratelimitedlog.RateLimitedLogFactory;
+import org.graylog2.shared.messageq.Acknowledgeable;
 import org.graylog2.shared.utilities.ExceptionUtils;
 import org.joda.time.DateTime;
 import org.slf4j.Logger;
