+++ graylog2-server/src/main/java/org/graylog2/shared/messageq/Acknowledgeable.java
@@ -14,7 +14,7 @@
  * along with this program. If not, see
  * <http://www.mongodb.com/licensing/server-side-public-license>.
  */
-package org.graylog2.indexer.messages;
+package org.graylog2.shared.messageq;
 
 import javax.annotation.Nullable;
